Refactoring, delete exception throw methods, add @NonNull

+++ src/main/java/faang/school/godbless/task/lru/cache/Data.java
@@ -8,12 +8,12 @@ import java.time.format.DateTimeFormatter;
 
 public class Data {
     @Getter
-    private final int id;
+    private final Long id;
     private String value;
     @Setter
     private LocalDateTime timeStamp;
 
-    public Data(int id, String value) {
+    public Data(Long id, String value) {
         this.id = id;
         this.value = value;
         this.timeStamp = LocalDateTime.now();
